change: move EntityClassGroupHelper method into WorldHelper
